Rename fancyWriter to httpFancyWriter and move NewWrapResponseWriter to pre 1.8 file
